+++ infrastructure/restapi/src/test/java/tech/pegasys/teku/infrastructure/restapi/RestApiTest.java
@@ -35,7 +35,7 @@ import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException
 class RestApiTest {
   private final Javalin app = mock(Javalin.class);
 
-  private final RestApi restApi = new RestApi(app, Optional.empty(), Optional.empty());
+  private final RestApi restApi = new RestApi(app, Optional.empty());
 
   @Test
   void start_shouldThrowInvalidConfigurationExceptionWhenPortInUse() {
@@ -55,16 +55,13 @@ class RestApiTest {
 
   @Test
   @DisabledOnOs(OS.WINDOWS)
-  void start_shouldFailFastWhenTokenNotWritable(@TempDir final Path tempDir) throws IOException {
-
+  void build_shouldFailFastWhenTokenNotWritable(@TempDir final Path tempDir) throws IOException {
     final Path managerDir = tempDir.resolve("manager");
     Files.createDirectory(
         managerDir,
         PosixFilePermissions.asFileAttribute(PosixFilePermissions.fromString("--x--x---")));
 
-    final RestApi restApi =
-        new RestApi(app, Optional.empty(), Optional.of(managerDir.resolve("pass")));
-    assertThatThrownBy(restApi::start).isInstanceOf(IllegalStateException.class);
-    assertThat(restApi.getRestApiDocs()).isEmpty();
+    assertThatThrownBy(() -> RestApiBuilder.ensurePasswordFile(managerDir.resolve("pass")))
+        .isInstanceOf(IllegalStateException.class);
   }
 }