+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Additions and Improvements
 
+- Implemented PostPtcDuties rest api endpoint (gloas api).
 - Added `/eth/v2/node/version` endpoint to retrieve structured version information for both beacon node and execution client.
 - Added deprecation warning on startup for any leveldb database types.
 - Increased default timeout of Engine API Get Payload requests to 2 seconds.